% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_iramuteq.R
\name{export_to_iramuteq}
\alias{export_to_iramuteq}
\title{Export a vector to an IRaMuTeQ formatted corpus file}
\usage{
export_to_iramuteq(
  x,
  dir = utils::choose.dir(),
  file_name = paste0(deparse(substitute(x)), ".txt"),
  var_name = "respondent"
)
}
\arguments{
\item{x}{An \link[base:is.recursive]{atomic vector} (usually a
\link[=character]{character} object) with a corpus (a collection of documents
containing text). Each element of \code{x} will represent a document (e.g., a
response from a survey).}

\item{dir}{(optional) a string indicating the directory on where the
corpus files must be saved (default: \code{utils::choose.dir()}).}

\item{file_name}{(optional) a string indicating the file name of the corpus.}

\item{var_name}{(optional) a string indicating the variable name for each
document. This is required by IRaMuTeQ for distinguish each document inside
a corpus (default: \code{"respondent"}).}
}
\value{
An invisible \code{NULL}. This function don't aim to return values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{export_to_iramuteq()} export atomic vectors to files with the format
required by IRaMuTeQ for text corpuses (a collection of documents containing
text).
}
\examples{
\dontrun{
if (requireNamespace("friends", quietly = TRUE)) {
    export_to_iramuteq(head(friends::friends[[1]], 100),
                       dir = tempdir(), file_name = "test.txt")
}
}
}
\concept{Utility functions}
