% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iramuteqlike-package.R
\docType{package}
\name{iramuteqlike-package}
\alias{iramuteqlike}
\alias{iramuteqlike-package}
\title{iramuteqlike: Tools to Reproduce the IRaMuTeQ Software Analyses}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Tools to reproduce the IRaMuTeQ software analyses in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipso.github.io/iramuteqlike/}
  \item \url{https://github.com/gipso/iramuteqlike/}
  \item Report bugs at \url{https://github.com/gipso/iramuteqlike/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID})

Authors:
\itemize{
  \item Alesandra Araujo de Souza \email{alesandra.araujo@mail.uft.edu.br} (\href{https://orcid.org/0000-0003-0303-169X}{ORCID})
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [copyright holder, funder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
