% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_word_frequency.R
\name{plot_word_frequency}
\alias{plot_word_frequency}
\title{Plot word frequencies}
\usage{
plot_word_frequency(
  data,
  min_freq = 1,
  max_words = 10,
  dist = FALSE,
  color = viridis::viridis(1)
)
}
\arguments{
\item{data}{A \code{\link[tibble:tibble]{tibble}} object with two columns: \code{word}
(with an unique set of words) and \code{freq} (with the absolute frequency of a
word in \code{word}). Use \code{\link[=word_frequency]{word_frequency()}} to generate this value.}

\item{min_freq}{(optional) a number indicating the minimum frequency that a
word can have to be plotted (default: \code{1}).}

\item{max_words}{(optional) a number indicating the maximum number of words
to be plotted (default: \code{100}).}

\item{dist}{(optional) a \code{\link[base:logical]{logical}} value
indicating if the function must plot a frequency \emph{versus} range chart
for visualizing the whole distribution (default: \code{FALSE}).}

\item{color}{(optional) a string with a Hex color code indicating the color
of the data (default: \code{viridis::viridis(1)}).}
}
\value{
An invisible \code{NULL}. This function don't aim to return values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{plot_word_frequency()} plots word frequencies charts related to an output of
\code{\link[=word_frequency]{word_frequency()}}.
}
\examples{
if (requireNamespace("friends", quietly = TRUE)) {
    data <- word_frequency(head(friends::friends[[1]], 100))
    plot_word_frequency(data)
    suppressMessages(plot_word_frequency(data, dist = TRUE))

    ## Note that graphics rendered in the documentation may
    ## not represent the quality of the output.
}

if (requireNamespace("stringi", quietly = TRUE)) {
    data <- word_frequency(stringi::stri_rand_lipsum(5))
    plot_word_frequency(data)
    suppressMessages(plot_word_frequency(data, dist = TRUE))

    ## Note that graphics rendered in the documentation may
    ## not represent the quality of the output.
}
}
\seealso{
Other Plot functions: 
\code{\link{plot_wordcloud}()}
}
\concept{Plot functions}
