% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wordcloud.R
\name{plot_wordcloud}
\alias{plot_wordcloud}
\title{Plot a word cloud}
\usage{
plot_wordcloud(
  data,
  min_freq = 2,
  max_words = 100,
  random_order = FALSE,
  rot_per = 0.35,
  colors = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{\link[tibble:tibble]{tibble}} object with two columns: \code{word}
(with an unique set of words) and \code{freq} (with the absolute frequency of a
word in \code{word}). Use \code{\link[=word_frequency]{word_frequency()}} to generate this value.}

\item{min_freq}{(optional) A number indicating the minimum frequency that a
word can have to be plotted (default: \code{2}).}

\item{max_words}{(optional) A number indicating the maximum number of words
to be plotted (default: \code{100}).}

\item{random_order}{(optional) A \code{\link[base:logical]{logical}} value
indicating if the function must plot words in random order. If \code{FALSE},
the words will be plotted in decreasing frequency. See
\link[wordcloud:wordcloud]{wordcloud()} to learn more (default: \code{FALSE}).}

\item{rot_per}{(optional) A number indicating the proportion words with 90
degree rotation (default: \code{0.35}).}

\item{colors}{(optional) A \code{\link[base:character]{character}} object with Hex
color codes indicating the color that the words must have, from least to
most frequent. If \code{NULL}, the \link[viridis:reexports]{viridis plasma palette}
will be used (default: \code{NULL}).}

\item{...}{(optional) Additional parameters to be passed to the
\link[wordcloud:wordcloud]{wordcloud()} function.}
}
\value{
An invisible \code{NULL}. This function don't aim to return values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{plot_wordcloud()} plots a word cloud using an output of \code{\link[=word_frequency]{word_frequency()}}.

This function uses \link[wordcloud:wordcloud]{wordcloud()} to generate
word cloud graphics.
}
\examples{
if (requireNamespace("friends", quietly = TRUE)) {
    data <- word_frequency(head(friends::friends[[1]], 100))
    plot_wordcloud(data)
}

if (requireNamespace("stringi", quietly = TRUE)) {
    data <- word_frequency(stringi::stri_rand_lipsum(5))
    plot_wordcloud(data)
}
}
\seealso{
Other Plot functions: 
\code{\link{plot_word_frequency}()}
}
\concept{Plot functions}
