% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iramuteqlike-package.R
\docType{package}
\name{iramuteqlike-package}
\alias{iramuteqlike}
\alias{iramuteqlike-package}
\title{iramuteqlike: Tools to Reproduce the IRaMuTeQ Software Analyzes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of tools to reproduce the IRaMuTeQ software analyzes on R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://giperbio.github.io/iramuteqlike/}
  \item \url{https://github.com/giperbio/iramuteqlike/}
  \item Report bugs at \url{https://github.com/giperbio/iramuteqlike/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

Authors:
\itemize{
  \item Alesandra Araujo de Souza \email{alesandra.araujo@mail.uft.edu.br} (\href{https://orcid.org/0000-0003-0303-169X}{ORCID})
}

}
\keyword{internal}
